Save restaurants info. Move functions.
